split assisted navigation to 2 node
